requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
fpdf2>=2.7.0
